requirements.txt
streamlit==1.28.1
python-dotenv==1.0.0